+++ src/pos_tui/widgets/__init__.py
@@ -4,6 +4,8 @@ Widgets for the POS Textual UI.
 This module contains custom widgets used throughout the POS application.
 """
 
+from .item_detail_modal import ItemDetailModal
+
 # Import widgets as they are implemented
 from .item_form import ItemEntryForm
 from .item_table import ItemTable
@@ -15,5 +17,6 @@ __all__: list[str] = [
     "ItemEntryForm",
     "ItemTable",
     "LinkTree",
+    "ItemDetailModal",
     # "Message",
 ]